The subclass files have been moved to the "Additional" folder. The "TalkWidget" is now transparent. Fixed the on click wake-up dialog.
